refactor(docs): date input dx
