+++ apps/docs/content/components/date-input/granularity.ts
@@ -1,60 +1,15 @@
-const App = `import {DateInput} from "@nextui-org/react";
-import {now, parseAbsoluteToLocal} from "@internationalized/date";
-import {useDateFormatter} from "@react-aria/i18n";
-
-export default function App() {
-  let [date, setDate] = React.useState(parseAbsoluteToLocal("2021-04-07T18:45:22Z"));
-
-  return (
-    <div className="w-full max-w-xl flex flex-col items-start gap-4">
-      <DateInput
-        granularity="second"
-        label="Date and time"
-        value={date}
-        onChange={setDate}
-      />
-      <DateInput granularity="day" label="Date" value={date} onChange={setDate} />
-      <DateInput granularity="second" label="Event date" />
-      <DateInput
-        granularity="second"
-        label="Event date"
-        placeholderValue={now("America/New_York")}
-      />
-    </div>
-  );
-}`;
-
-const AppTs = `import {DateInput} from "@nextui-org/react";
-import {DateValue, now, parseAbsoluteToLocal} from "@internationalized/date";
-import {useDateFormatter} from "@react-aria/i18n";
-
-export default function App() {
-  let [date, setDate] = React.useState<DateValue>(parseAbsoluteToLocal("2021-04-07T18:45:22Z"));
-
-  return (
-    <div className="w-full max-w-xl flex flex-col items-start gap-4">
-      <DateInput
-        granularity="second"
-        label="Date and time"
-        value={date}
-        onChange={setDate}
-      />
-      <DateInput granularity="day" label="Date" value={date} onChange={setDate} />
-      <DateInput granularity="second" label="Event date" />
-      <DateInput
-        granularity="second"
-        label="Event date"
-        placeholderValue={now("America/New_York")}
-      />
-    </div>
-  );
-}`;
+import App from "./granularity.raw.jsx?raw";
+import AppTs from "./granularity.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
+};
+
+const reactTs = {
   "/App.tsx": AppTs,
 };
 
 export default {
   ...react,
+  ...reactTs,
 };
