+++ apps/docs/content/components/date-input/label-placements.ts
@@ -1,26 +1,4 @@
-const App = `import {DateInput} from "@nextui-org/react";
-import {CalendarDate} from "@internationalized/date";
-
-export default function App() {
-  const placements = [
-    "inside",
-    "outside",
-    "outside-left",
-  ];
-
-  return (
-    <div className="w-full flex flex-col max-w-sm gap-4">
-      {placements.map((placement) => (
-        <DateInput 
-          label="Birth date" 
-          placeholderValue={new CalendarDate(1995, 11, 6)} 
-          description={placement}
-          labelPlacement={placement}
-        />
-      ))}
-    </div>  
-  );
-}`;
+import App from "./label-placements.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
